Feat: Reorganizar estrutura de componentes, movendo Home para a pasta pages e atualizando importações; atualizar mensagem de copyright no Footer
