+++ src/App.tsx
@@ -1,5 +1,5 @@
 
-import Home from './components/Home'
+import Home from './pages/Home'
 function App() {
   
   return (
